requirements.txt
flask==3.0.2
pandas==2.2.0
pandasai==2.0.4
openai==1.12.0
python-dotenv==1.0.1
matplotlib==3.8.2
gunicorn==21.2.0 